feat: use useRequireBluetooth hook and drawer in SkipSelectDevice component
